+++ src/onegov.wtfs/onegov/wtfs/assets/js/common.js
@@ -133,3 +133,9 @@ Intercooler.ready(function(element) {
 
     processCommonNodes(el, true);
 });
+
+
+var updateDispatchDates = function(dispatch_date, return_date) {
+    $('#dispatch_date').val(dispatch_date);
+    $('#return_date').val(return_date);
+}